requirements.txt
asgiref==3.6.0
boto3==1.15.3
botocore==1.18.18
dj-database-url==1.3.0
Django==3.2.16
django-auto-logout==0.5.1
django-crispy-forms==2.0
django-phonenumber-field==7.0.2
gunicorn==20.1.0
jmespath==0.10.0
phonenumbers==8.13.8
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
s3transfer==0.3.7
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.25.11
whitenoise==6.4.0
